Move events back up to core, because moving it to a separate folder was a mistake.
