--- core/events/variable/rename.js
@@ -1,91 +1,0 @@
-/**
- * @license
- * Visual Blocks Editor
- *
- * Copyright 2018 Google Inc.
- * https://developers.google.com/blockly/
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * @fileoverview Variable rename events fired as a result of actions in
- *     Blockly's editor.
- * @author fraser@google.com (Neil Fraser)
- */
-'use strict';
-
-goog.provide('Blockly.Events.VarRename');
-
-goog.require('Blockly.Events');
-goog.require('Blockly.Events.VarBase');
-goog.require('goog.array');
-goog.require('goog.math.Coordinate');
-
-/**
- * Class for a variable rename event.
- * @param {Blockly.VariableModel} variable The renamed variable.
- *     Null for a blank event.
- * @param {string} newName The new name the variable will be changed to.
- * @extends {Blockly.Events.VarBase}
- * @constructor
- */
-Blockly.Events.VarRename = function(variable, newName) {
-  if (!variable) {
-    return;  // Blank event to be populated by fromJson.
-  }
-  Blockly.Events.VarRename.superClass_.constructor.call(this, variable);
-  this.oldName = variable.name;
-  this.newName = newName;
-};
-goog.inherits(Blockly.Events.VarRename, Blockly.Events.VarBase);
-
-/**
- * Type of this event.
- * @type {string}
- */
-Blockly.Events.VarRename.prototype.type = Blockly.Events.VAR_RENAME;
-
-/**
- * Encode the event as JSON.
- * @return {!Object} JSON representation.
- */
-Blockly.Events.VarRename.prototype.toJson = function() {
-  var json = Blockly.Events.VarRename.superClass_.toJson.call(this);
-  json['oldName'] = this.oldName;
-  json['newName'] = this.newName;
-  return json;
-};
-
-/**
- * Decode the JSON event.
- * @param {!Object} json JSON representation.
- */
-Blockly.Events.VarRename.prototype.fromJson = function(json) {
-  Blockly.Events.VarRename.superClass_.fromJson.call(this, json);
-  this.oldName = json['oldName'];
-  this.newName = json['newName'];
-};
-
-/**
- * Run a variable rename event.
- * @param {boolean} forward True if run forward, false if run backward (undo).
- */
-Blockly.Events.VarRename.prototype.run = function(forward) {
-  var workspace = this.getEventWorkspace_();
-  if (forward) {
-    workspace.renameVariableById(this.varId, this.newName);
-  } else {
-    workspace.renameVariableById(this.varId, this.oldName);
-  }
-};
