+++ core/events_abstract.js
